feat: rename from `Clever` to `Qlever`, thus matching the brand-new logo
